+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Unreleased in the current development version (target v0.23.0):
 - Histogram: integration in AQUA analysis (#85)
 
 Complete list:
+- Global Biases: add mean value and RMSE to global bias plot (#132)
 - Teleconnections: fix timmean assignment in ENSO, NAO, MJO diagnostics (#129)
 - Seaice: Add gridlines in maps (#125)
 - Add cross-check workflow for testing against aqua-core main or tag/branch (#126)